chore: move fonts to utils root
